security and role added

+++ src/main/java/com/studymate/NotesApplication/services/EntryService.java
@@ -6,6 +6,7 @@ import com.studymate.NotesApplication.repository.todoRepo;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -19,8 +20,9 @@ public class EntryService {
     @Autowired
     private UserService userService;
 
-    public void addTodoService(NewEntry newEntry , ObjectId id){
-        User user = userService.findUserById(id)
+    @Transactional
+    public void addTodoService(NewEntry newEntry , String username){
+        User user = userService.findUserByUsername(username)
                 .orElseThrow(() -> new RuntimeException("User not found"));
         newEntry.setDate(LocalDateTime.now());
         NewEntry saved =  newtodo.save(newEntry);
@@ -28,12 +30,12 @@ public class EntryService {
         userService.saveUser(user);
     }
 
-    public void addTodoService(NewEntry newEntry){
+    public void upDateTodoService(NewEntry newEntry){
         newtodo.save(newEntry);
     }
 
-    public List<NewEntry> getAlltoDosByUser(ObjectId id){
-        User user = userService.findUserById(id).orElseThrow(() ->new  RuntimeException("User not found"));
+    public List<NewEntry> getAlltoDosByUser(String  username){
+        User user = userService.findUserByUsername(username).orElseThrow(() ->new  RuntimeException("User not found"));
        return user.getUserEntries();
 //        return newtodo.findAll();
     }
@@ -42,12 +44,18 @@ public class EntryService {
        return newtodo.findById(id);
     }
 
-    public boolean deletebyIdService(ObjectId id, ObjectId userID){
-        User user = userService.findUserById(userID).orElseThrow(()-> new RuntimeException("User not found"));
-        user.getUserEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveUser(user);
-        newtodo.deleteById(id);
-        return true;
+    @Transactional
+    public boolean deletebyIdService(ObjectId id, String username){
+        User user = userService.findUserByUsername(username).orElseThrow(()-> new RuntimeException("User not found"));
+        boolean check = user.getUserEntries().removeIf(x -> x.getId().equals(id));
+        if(check) {
+            userService.saveUser(user);
+            newtodo.deleteById(id);
+            return true;
+        }
+        else {
+            return false;
+        }
     }
 
 }
